pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=65.5.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cron-mng"
version = "0.1.0"
description = "A CLI application that orchestrates cron jobs"
readme = "README.md"
authors = [{ name = "Pavat Pichetsin", email = "pavatpichetsin@outlook.com" }]
dependencies = [
    "click >= 8.1.3",
    "python-crontab >= 3.1.0"
]

[project.scripts]
cron-mng = "app.__main__:main"